updated key in deploy.yml
Siqi, please note this change. I think you have "instance" instead of "Ninstances" on your branch (which is why 'instance' keeps popping out on master, causing the deployment to crap out from time to time...)

Let's make sure all the deployment/cluster etc. yml files have the key "Ninstances" in them instead of 'instance' because it is more descriptive of what the input is supposed to be.